fixed lines

+++ sgl/src/sgl_core.cpp
@@ -131,17 +131,17 @@ void SglCore::draw_circle(SglVertex center, float radius) {
     // SGL_DEBUG_OUT("[SglCore::push_vertex()] vertex transformed by modelview and projection: \n" + vertex.to_string());
     center = contexts.at(current_context).viewport_mat * center;
 
-    renderer.draw_circle(center, radius);
+    renderer.draw_circle(center, static_cast<int>(radius));
 }
 
 
 void SglCore::draw_ellipse(SglVertex center, float a, float b) {
     //add transforms
-    renderer.draw_ellipse(center, a, b);
+    renderer.draw_ellipse(center, static_cast<int>(a), static_cast<int>(b));
 }
 
 
 void SglCore::draw_arc(SglVertex center, float radius, float from, float to) {
     //add transforms
-    renderer.draw_arc(center, radius, from, to);
+    renderer.draw_arc(center, static_cast<int>(radius), static_cast<int>(from), static_cast<int>(to));
 }
